refactor check tool, move ack kafka writer unit test to use the new check tool
